+++ config/global.ini.php
@@ -226,20 +226,6 @@ enabled_periods_UI = "day,week,month,year,range"
 ; month periods are disabled, then year periods can no longer be computed, so they are effectively disabled as well.
 enabled_periods_API = "day,week,month,year,range"
 
-; whether to enable segment archiving cache
-; Note: if you use any plugins, this need to be compliant with Matomo and
-; * depending on the segment you create you may need a newer MySQL version (eg 5.7 or newer)
-; * use a reader database for archiving in case you have configured a database reader
-enable_segments_cache = 1
-
-; whether to enable subquery cache for Custom Segment archiving queries
-enable_segments_subquery_cache = 0
-; Any segment subquery that matches more than segments_subquery_cache_limit IDs will not be cached,
-; and the original subquery executed instead.
-segments_subquery_cache_limit  = 100000
-; TTL: Time to live for cache files, in seconds. Default to 60 minutes
-segments_subquery_cache_ttl  = 3600
-
 ; when set to 1, all requests to Matomo will return a maintenance message without connecting to the DB
 ; this is useful when upgrading using the shell command, to prevent other users from accessing the UI while Upgrade is in progress
 maintenance_mode = 0